Make the command-line interface behave more like `tsc -b`
Previously, lsif-typescript had unexpected behavior for how it
interpreted command-line flags. For example, it was not possible to
pass it a custom tsconfig.json path. It was also not possible to provide
a list of tsconfig.json files, it was only possible to index a list of
projects when using Yarn workspaces. On top of this, if you passed
incorrectly typed command-line flags then they got silently ignored due
to the parser we were using (yargs).

This commit fixes most of these quirks:

- lsif-typescript index now accepts a repeated list of tsconfig
  references, using the same syntax as `tsc -b`.
- lsif-typescript now reports errors for incorrect command-line flags.

Fixes #48