fix: Use allowJS in --infer-tsconfig. (#83)
lsif-node uses allowJS; however, that was skipped for some reason
in lsif-typescript. Without allowJS, JS files will be excluded
by the TypeScript compiler.

The test is intended for future iteration; and the correctness
of the results doesn't really matter for this PR. The key part
is that we are able to generate LSIF dumps properly for JS.

Fixes issue #81.