Update `--infer-tsconfig` to skip the `allowJs` option when there exist TS files (#90)
Fixes #87. Previously, we always included the `allowJs` option when
using `--infer-tsconfig`. However, this behavior was undesirable because
TypeScript package repos include a large amount of generated JavaScript
code that should not be indexed.  This commit changes the behavior so
that we only use `allowJs` when the project doesn't have any TypeScript
files.

For example, the `npm/typescript` package (for the TypeScript compiler
itself) times out when we index with `allowJs` while it completes in
a few seconds without `allowJs`.  It's OK to skip the JavaScript files
since the TypeScript files include the definitions for the JavaScript
code.