Add tests for command-line parser (#102)
* Add tests for command-line parser

Previously, we didn't have any tests for how we parsed command-line
flags. This was problematic because the tests didn't catch breaking
changes in seemingly innocent pull requests like https://github.com/sourcegraph/lsif-typescript/pull/100
that rename a variable. Now, we have tests for the command-line parser
so that we hopefully catch more unintentional regressions in the future.

* Step 1: rename infertsconfig file to fix capitalization

* Step 2: rename infertsconfig file back with correct capitalization

* Fix parsing of --cwd and address review comments